tdf#101873 sw: fix search for second term after double not found
Regression from commit 79060186c9fff38e5744ce5e1fa007fdecd6b7fe
(SwView::SearchAndWrap: fix WrapAround search in fly frames,
2015-06-03).

The search implementation looks in the body text, then in case searching
in extra content was not enabled already, it attempts to search there.

The problem was that while the wrap-around implementation intended to do
the same, it failed to unset the "search in extra" flag in case nothing
was found, fix that now.

Change-Id: Idf078e464824cad69c9709309435f2db207503b8
Reviewed-on: https://gerrit.libreoffice.org/67986
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>