oox smartart, cycle matrix: handle destination order in connections
It is possible to have connections from multiple data nodes to the same
presentation node with a presOf type. We use to order these based on as
they appear in the data XML, but we need to order them according to the
destOrd attribute.

Introduce an std::map for that, so get ordering automatically as we
iterate. Turn the std::pair into a struct to make the code a bit more
readable.

Change-Id: I3d2bb047ed3f171a194851f89151bd94071a8176
Reviewed-on: https://gerrit.libreoffice.org/68027
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins