tdf#122582: Add touch gestures to iOS app slideshow, also for exiting it
Left swipt shows the next slide, right swipe shows the previoius
slide. Up or down swipe exit the slideshow.

Use the same hammer.js as Online does for touch gestures in dialogs.
With one addition that was necessary to avoid a runtime error: look
for a comment starting with '// tml:'. No idea why this is not needed
in Online.

Change-Id: I7a22b16dfbc069dcdc1be344005ab7503dcd1ba7