tdf#123393 RTF import: fix too big font size in table cell
The reason was that A2 had an explicit paragraph style reference, but A1
did not, so table buffering caused A2 style to affect A1 style as well.

Combine this with style deduplication, and then A2 style considered the
direct formatting (font size) in A1 to be redundant, so it was lost on
import.

Fix the problem by moving the copy&pasted properties buffering to a
single function, and there buffering not only the properties, but also
the active style index.

Change-Id: I99f2020b8bef237849fd622b25ac5ef0516d69e4
Reviewed-on: https://gerrit.libreoffice.org/68361
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>