oox smartart, picture strip: fix lack of spacing around the picture list
The snake algorithm in PowerPoint seem to interpret spacing as follows:
if you have N elements, then there should be the requested amount of
spacing between the elements, and also double amount of spacing around
the actual list of elements.

With this, the SmartArt and the title shape in the bugdoc no longer
overlaps.

Change-Id: I5d6885b434bfaff9de9aac595a298a5346524e19
Reviewed-on: https://gerrit.libreoffice.org/68397
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins