Auto-redact - First stab
Add a SfxRedactionHelper::autoRedactPage method which searches for the given term
through the gdimetafile which has the whole content of an xPage (a Draw page),
and draws redaction rectangles at proper positions with proper sizes.

The search is case sensitive, and finds only the first occurences on a line.
Will switch to a proper search provider via a follow-up patch.

Change-Id: If3db62e50994670143785b6727fdcf1ccd4c6f8e
Reviewed-on: https://gerrit.libreoffice.org/68597
Tested-by: Jenkins
Reviewed-by: Muhammet Kara <muhammet.kara@collabora.com>