tdf#123446 sw_redlinehide: fix crash on Redo of ToX
The problem is that the ToX must be updated with the same layout redline
setting as it was originally created, so that subsequent Redo actions
see the expected node indexes.

Unfortunately it's not enough to just pass a flag to the ToX update
functions, because they check GetTextNodeForParaProps() so we need a
layout corresponding to the layout setting; if there isn't one, the
existing one is temporarily toggled.

This could be much better if the MergedPara would be independent of the
layout and always exist, but with the various SwModify design issues
that looks tricky to do...

(regression from e5f95e6a742ea7f75b9d9b131731fc6d03949d1f)

Change-Id: Ibdc5b4ace54ace27e5223a25ecaf39bb493fb69b
Reviewed-on: https://gerrit.libreoffice.org/68704
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>
Tested-by: Michael Stahl <Michael.Stahl@cib.de>