tdf#123819: Revert "Use comphelper::getProcessComponentContext()"
Even though in theory it was worse to get an exception and return an
empty Any for CommandBars, in practice there are not resources to
figure out why that change apparently then caused a crash in the
customer application. Let's hope that it doesn't actually need the
CommandBars object for anything essential.

This reverts commit 87b4bd61792b28fe475c71d4484cd219c1e533ae.

Change-Id: Id2ec3a7254c5dd5a36ce2d5e930a457b9abd44f1
Reviewed-on: https://gerrit.libreoffice.org/68881
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>