tdf#121120 vcl opengl: handle 8-bit palette bitmap as 24 bit RGB texture
This fixes the black rectangle instead of actual image content in the
bugdoc.

An alternative would be to handle this in the PNG import, but commit
087936b69f627c97fa427f6ba8acc41f797b4e70 (tdf#113918: Workaround: Load
1bpp indexed PNG as 8bpp indexed Bitmap, 2017-11-20) already tried
something similar and failed, leading to the revert in commit
beff39f36befa5e420a04fa8a825c2d2606cd181 (tdf#115297: alternative fix
for displaying 1bit images, 2018-02-27).

The test is especially useful with SAL_FORCEGL=1 SAL_USE_VCLPLUGIN=gen
specified on Linux, so the GL codepath is triggered.

Change-Id: Ia9d049d09dce2ac34826ee427f74616a96a35c88
Reviewed-on: https://gerrit.libreoffice.org/68925
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins