tdf#121092: save GenerateASBeforeCorrelationName to ODB
As mentioned in connectivity/registry/README, the property should
be named differently in Features and Properties sections.

Also the default value set in ODbAdminDialog::createItemSet should
be synchronized with the other defaults, because otherwise, in case
when the value is absent in the set (when it is default), this value
is used in the dialog... sigh, we have too much places where relevant
defaults are generated and need to be in sync. TODO: this should be
reimplemented to take all defaults from the xcu.

Change-Id: Ia61ab2e585107ede9c977cf563cb30e45d624bed
Reviewed-on: https://gerrit.libreoffice.org/69002
Tested-by: Jenkins
Reviewed-by: Lionel Elie Mamane <lionel@mamane.lu>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>