writerfilter: implement RTF derived styles defaulting
It turns out that the situation fixed in commit
0016cd3ccb955a7b5188a4e0e9559f871d95cf25 also applies to the definition
of the styles themselves.

To implement the same style import as Word, the style definitions need
to be stored twice: once as read from the file, and another time with
attributes defaulted and deduplicated vs. the parent style; the second
representation is then sent to the domain mapper.

To make this easier, add a bool parameter to cloneAndDeduplicate()
to disable the implicit pPr dereferencing that happens when creating the
hard formatted paragraph properties (this could potentially be cleaned
up further if those paragraph properties would use pPr wrapper
themselves).

Also implement defaulting of line spacing in getDefaultSPRM().

Change-Id: I4810e917697b3af244e5dbdd7f5a45b4767c93fc
Reviewed-on: https://gerrit.libreoffice.org/70320
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>