tdf#124327 sc opengl: fix lack of AA text in charts
Commit d6b3d48a248f4d6c0c56c93e658a1541831d8f29 (sc lok: emit
RowColumnHeader info in twips, 2015-11-02) and later commits improved
precision of the Pixel Per Twips (PPT) factor in Calc, but this caused
offset problems in PDF export. So the LOK case uses double precision and
otherwise we use limited precision since commit
e295a06a9ead9ed3a094a8f25295e9cfbe9ed748 (tdf#115262 sc: fix cumulative
placement error of images, 2018-10-01).

It turns out this double precision was also useful in the opengl case:
if UI scaling is set to 125% on Windows, then this helps producing AA
text for charts. Given that unlimited precision is the direction we want
to go, enable double precision for the GL case as well, not just for
LOK.

Verified that tdf#115262 is not re-introduced in the GL case, also that
manual PDF export (with GL enabled) does not produce the mentioned
unexpected image position offsets.

Change-Id: I23a8d6dce1b48c178977bda9909feb59fe8f3ad8
Reviewed-on: https://gerrit.libreoffice.org/70407
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins