gtk3_kde5: No longer kill fpicker process on timeout
In some observed cases, 100 ms were not enough for the separate
'lo_kde5filepicker' process to terminate properly, in particular
on slow or busy hardware.
Drop the timeout here and just wait for the process to finish
to avoid killing it unnecessarily, which could e.g. lead to file
dialog settings/state not properly being saved by the native KDE
Plasma QFileDialog (plasma-integration, e.g. in
'KDEPlatformFileDialogHelper::saveSize'), so that the last
used mode or size would not be remembered.

The assumption is that the separate fpicker process should now
properly finish in all cases, in particular since commit
13e1e373757297cbae543695ff0e4e7c06b4e2a4.

Should that assumption not hold, let's rather fix the underlying
issues than increasing the timeout here.

Change-Id: I0a58c1425940fbfaecb5012a24fd9a0222cbaccb
Reviewed-on: https://gerrit.libreoffice.org/70834
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>