tdf#109376 sw: fix redline SwUndoDelete with end pos on SwTableNode crash
... that happens when you accept a delete redline (or reject an insert
redline) with such end pos.

The problem is that first a DeleteRange() will move the anchor position
onto the table node (because check in SwUndoSaveContent::DelContentIndex()
is surprisingly asymmetric and so the fly not deleted by the previous
bugfix), then DelFullPara() creates a second SwUndoDelete then deleting
the fly crashes because its anchors was moved.

The code in lcl_AcceptRedline() / lcl_RejectRedline() doesn't make much
sense (but always was like this), if we just call DeleteFullPara() once
instead, the problem is avoided, and we don't even have to worry about
why DelContentIndex() is so asymmetric (is "selection direction"
really a meaningful concept?).

Reportedly this started to crash with commit
60e78a788c481e157eb90cce7833855465f64b40, previously it was just wrong.

Change-Id: Ib3d4b31e0255a6f4e7b49b40f204dec168ea3006
Reviewed-on: https://gerrit.libreoffice.org/70836
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>